+++ .gitignore
@@ -1,3 +1,4 @@
+demo
 demo.toml
 *.log
 *.bak
