+++ CHANGELOG.md
@@ -1,5 +1,16 @@
 # 更新日志
 
+25w04a
+---
+- PRE-RELEASE: 此版本是v2的候选版本(技术验证版),请勿在生产环境中使用; 我们可能会撤除v2更新计划(若技术验证版顺利通过, 则会发布v2正式版)
+- CHANGE: 大幅修改核心组件
+
+1.8.2
+---
+- RELEASE: v1.8.2正式版发布; 这或许会是v1的最后一个版本
+- FIX: 修复部分日志表述错误
+- CHANGE: 关闭`gin`框架的`fmt`日志打印, 在高并发场景下提升一定性能(go 打印终端日志性能较差，可能造成性能瓶颈)
+
 25w03a
 ---
 - PRE-RELEASE: 此版本是v1.8.2的候选预发布版本,请勿在生产环境中使用
